Moved logic for parsing to its own class
